pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "english-verb-trainer"
version = "0.1.0"
description = "English Irregular Verb Trainer — DevOps Edition"
requires-python = ">=3.10"
dependencies = [
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
  "python-dotenv>=1.0",
  "typer>=0.12",
]

[project.scripts]
verb-trainer = "main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
